LOGIC 6 VERSION BASE PROJECT - Create admin - Create user not admin

+++ src/main/java/com/labotec/traccar/app/usecase/ports/input/repository/BusStopRepository.java
@@ -3,6 +3,6 @@ package com.labotec.traccar.app.usecase.ports.input.repository;
 import com.labotec.traccar.app.utils.common.repository.GenericRepository;
 import com.labotec.traccar.domain.database.models.BusStop;
 
-public interface BusStopRepository extends GenericRepository<BusStop, Integer> {
+public interface BusStopRepository extends GenericRepository<BusStop,Long> {
 
 }
