+++ src/main/java/com/labotec/traccar/app/usecase/ports/input/repository/DriverRepository.java
@@ -3,5 +3,5 @@ package com.labotec.traccar.app.usecase.ports.input.repository;
 import com.labotec.traccar.app.utils.common.repository.GenericRepository;
 import com.labotec.traccar.domain.database.models.Driver;
 
-public interface DriverRepository extends GenericRepository<Driver, Integer> {
+public interface DriverRepository extends GenericRepository<Driver,Long>  {
 }
