+++ src/main/java/com/labotec/traccar/app/usecase/ports/input/repository/RouteRepository.java
@@ -3,6 +3,6 @@ package com.labotec.traccar.app.usecase.ports.input.repository;
 import com.labotec.traccar.app.utils.common.repository.GenericRepository;
 import com.labotec.traccar.domain.database.models.Route;
 
-public interface RouteRepository extends GenericRepository<Route,  Integer> {
+public interface RouteRepository extends GenericRepository<Route,  Long> {
 
 }
