+++ src/main/java/com/labotec/traccar/app/usecase/ports/input/repository/VehicleRepository.java
@@ -3,7 +3,7 @@ package com.labotec.traccar.app.usecase.ports.input.repository;
 import com.labotec.traccar.app.utils.common.repository.GenericRepository;
 import com.labotec.traccar.domain.database.models.Vehicle;
 
-public interface VehicleRepository extends GenericRepository<Vehicle,Integer> {
+public interface VehicleRepository extends GenericRepository<Vehicle,Long> {
 
-    Vehicle findByDevice(Integer id);
+    Vehicle findByDevice(Long id);
 }
